+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### :wrench: Changed
 
 - Replace archived `backoff` with `tenacity`. [#3286](https://github.com/schemathesis/schemathesis/issues/3286)
+- Suppress Hypothesis reproduction blocks in Schemathesis pytest failures.
 
 ## [4.4.3](https://github.com/schemathesis/schemathesis/compare/v4.4.2...v4.4.3) - 2025-11-07